cras: Fix integer overflows in cras_shm bounds checks
Cast buffer_offset to uint64_t before performing addition in bounds
checks within cras_shm.h.
Previously, buffer_offset (32-bit unsigned) was added to other 32-bit
offsets/sizes (used_size, read_offset, write_offset) and compared
against samples_info.length (64-bit size_t). A malicious client could
provide a large offset to trigger a 32-bit integer overflow, wrapping
the sum to a small value. This small value would bypass the bounds check
against the 64-bit length, leading to an out-of-bounds write.
By casting buffer_offset to size_t, the addition is performed in 64-bit,
preventing the overflow and correctly rejecting out-of-bounds offsets.
This fixes the vulnerability in:
- cras_shm_buff_for_idx
- cras_shm_get_checked_read_offset
- cras_shm_get_checked_write_offset

BUG=b:515786674
TEST=bazel test //cras/src/tests:shm_unittest

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/7977362
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Auto-Submit: Yu-Hsuan Hsu <yuhsuan@chromium.org>

+++ cras/include/cras_shm.h
@@ -204,7 +204,8 @@ static inline uint8_t* cras_shm_buff_for_idx(const struct cras_audio_shm* shm,
   idx = idx & CRAS_SHM_BUFFERS_MASK;
 
   unsigned buffer_offset = cras_shm_get_checked_buffer_offset(shm, idx);
-  if (buffer_offset + shm->config.used_size > shm->samples_info.length) {
+  if ((uint64_t)buffer_offset + shm->config.used_size >
+      shm->samples_info.length) {
     buffer_offset = idx * shm->config.used_size;
   }
 
@@ -225,7 +226,7 @@ static inline unsigned cras_shm_get_checked_read_offset(
   if (read_offset > shm->config.used_size) {
     return 0;
   }
-  if (buffer_offset + read_offset > shm->samples_info.length) {
+  if ((uint64_t)buffer_offset + read_offset > shm->samples_info.length) {
     return 0;
   }
   return read_offset;
@@ -246,7 +247,7 @@ static inline unsigned cras_shm_get_checked_write_offset(
 
   /* If the buffer offset plus the write offset overruns the samples area,
    * return the longest valid write_offset */
-  if (buffer_offset + write_offset > shm->samples_info.length) {
+  if ((uint64_t)buffer_offset + write_offset > shm->samples_info.length) {
     return shm->samples_info.length - buffer_offset;
   }
   return write_offset;
